Module dprint_core::configuration

Structs

ConfigurationDiagnostic
Represents a problem within the configuration.

DefaultGlobalConfiguration

GlobalConfiguration

ParseConfigurationError

ResolveConfigurationResult

Enums

NewLineKind

Constants

DEFAULT_GLOBAL_CONFIGURATION

Functions

get_value
If the provided key exists, takes its value from the provided config and returns it. If the provided key does not exist, it returns the default value. Adds a diagnostic if there is any problem deserializing the value.

resolve_global_config
Resolves a collection of key value pairs to a GlobalConfiguration.

resolve_new_line_kind
Resolves the NewLineKind text from the provided file text and NewLineKind.