Function dprint_core::configuration::get_value

pub fn get_value<T>(
    config: &mut ConfigKeyMap, 
    key: &'static str, 
    default_value: T, 
    diagnostics: &mut Vec<ConfigurationDiagnostic>
) -> T where
    T: FromStr,
    <T as FromStr>::Err: Display, 

If the provided key exists, takes its value from the provided config and returns it. If the provided key does not exist, it returns the default value. Adds a diagnostic if there is any problem deserializing the value.