Module dprint_core::formatting::conditions

Structs

NewLineIfHangingSpaceOtherwiseOptions

Functions

force_reevaluation_once_resolved

This condition can be used to force the printer to jump back to the point this condition exists at once the provided info is resolved.

if_above_width

Prints the provided items when the current relative column number is above the specified width.

if_above_width_or

Prints the provided true_items when the current relative column number is above the specified width or prints the false_items otherwise.

if_false

if_true

if_true_or

indent_if_start_of_line

new_line_if_hanging

new_line_if_hanging_space_otherwise

new_line_if_multiple_lines_space_or_new_line_otherwise

single_indent_if_start_of_line

with_indent_if_start_of_line_indented