Enum dprint_core::formatting::Signal

pub enum Signal {

    NewLine,
    Tab,
    PossibleNewLine,
    SpaceOrNewLine,
    ExpectNewLine,
    QueueStartIndent,
    StartIndent,
    FinishIndent,
    StartNewLineGroup,
    FinishNewLineGroup,
    SingleIndent,
    StartIgnoringIndent,
    FinishIgnoringIndent,
    StartForceNoNewLines,
    FinishForceNoNewLines,
    SpaceIfNotTrailing,

}

Variants

NewLine

Signal that a new line should occur based on the printer settings.

Tab

Signal that a tab should occur based on the printer settings.

PossibleNewLine

Signal that the current location could be a newline when exceeding the line width.

SpaceOrNewLine

Signal that the current location should be a space, but could be a newline if exceeding the line width.

ExpectNewLine

Expect the next character to be a newline. If it’s not, force a newline.

QueueStartIndent

Queue a start indent to be set after the next written item.

StartIndent

Signal the start of a section that should be indented.

FinishIndent

Signal the end of a section that should be indented.

StartNewLineGroup

Signal the start of a group of print items that have a lower precedence for being broken up with a newline for exceeding the line width.

FinishNewLineGroup

Signal the end of a newline group.

SingleIndent

Signal that a single indent should occur based on the printer settings.

StartIgnoringIndent

Signal to the printer that it should stop using indentation.

FinishIgnoringIndent

Signal to the printer that it should start using indentation again.

StartForceNoNewLines

Signal to the printer that it shouldn’t print any new lines.

FinishForceNoNewLines

Signal to the printer that it should finish not printing any new lines.

SpaceIfNotTrailing

Signal that a space should occur if not trailing.