Struct dprint_core::formatting::ConditionResolverContext

pub struct ConditionResolverContext<'a, 'b> {
    pub writer_info: WriterInfo,
    // some fields omitted
}

Context used when resolving a condition.

Fields
writer_info: WriterInfo

Gets the writer info at the condition’s location.

Implementations

impl<'a, 'b> ConditionResolverContext<'a, 'b>

pub fn get_resolved_condition(
&mut self,
condition_reference: &ConditionReference
) -> Option<bool>

Gets if a condition was true, false, or returns undefined when not yet resolved. A condition reference can be retrieved by calling the get_reference() on a condition.

pub fn get_resolved_info(&self, info: &Info) -> Option<&WriterInfo>

Gets the writer info at a specified info or returns undefined when not yet resolved.

pub fn clear_info(&mut self, info: &Info)

Clears the info result from being stored.

pub fn has_info_moved(&mut self, info: &Info) -> Option<bool>

Gets if the provided info has moved positions since the last check. Returns None when the info can’t be resolved. Returns Some(false) the first time this is called.