Struct dprint_core::Condition

pub struct Condition<TString = String, TInfo = Info> where
    TString: StringTrait,
    TInfo: InfoTrait,  {
    pub condition: Rc<Box<ConditionResolver<TString, TInfo, Condition<TString, TInfo>>>>,
    pub true_path: Option<PrintItemPath<TString, TInfo, Condition<TString, TInfo>>>,
    pub false_path: Option<PrintItemPath<TString, TInfo, Condition<TString, TInfo>>>,
    // some fields omitted
}

Conditionally print items based on a condition.

These conditions are extremely flexible and can even be resolved based on information found later on in the file.

Fields
condition: Rc<Box<ConditionResolver<TString, TInfo, Condition<TString, TInfo>>>>

The condition to resolve.

true_path: Option<PrintItemPath<TString, TInfo, Condition<TString, TInfo>>>

The items to print when the condition is true.

false_path: Option<PrintItemPath<TString, TInfo, Condition<TString, TInfo>>>

The items to print when the condition is false or undefined (not yet resolved).

Methods

impl<TString, TInfo> Condition<TString, TInfo> where
TString: StringTrait,
TInfo: InfoTrait,

pub fn new(
name: &'static str,
properties: ConditionProperties<TString, TInfo>
) -> Condition<TString, TInfo>

pub fn new_true() -> Condition<TString, TInfo>

pub fn new_false() -> Condition<TString, TInfo>

pub fn new_with_dependent_infos(
name: &'static str,
properties: ConditionProperties<TString, TInfo>,
dependent_infos: Vec<TInfo>
) -> Condition<TString, TInfo>

pub fn get_reference(&mut self) -> ConditionReference