Module dprint_core::configuration

Structs

ConfigurationDiagnostic

Represents a problem within the configuration.

DefaultGlobalConfiguration

GlobalConfiguration

ParseConfigurationError

ResolveConfigurationResult

Enums

ConfigKeyValue

NewLineKind

Constants

DEFAULT_GLOBAL_CONFIGURATION

Functions

get_nullable_value

If the provided key exists, takes its value from the provided config and returns it. If the provided key does not exist, it returns None. Adds a diagnostic if there is any problem deserializing the value.

get_unknown_property_diagnostics

Gets a diagnostic for each remaining key value pair in the hash map.

get_value

If the provided key exists, takes its value from the provided config and returns it. If the provided key does not exist, it returns the default value. Adds a diagnostic if there is any problem deserializing the value.

handle_renamed_config_property

If it exists, moves over the configuration value over from the old key to the new key and adds a diagnostic.

parse_config_key_map

Creates a ConfigKeyMap from a series of string key value pairs.

resolve_global_config

Resolves a collection of key value pairs to a GlobalConfiguration.

resolve_new_line_kind

Resolves the NewLineKind text from the provided file text and NewLineKind.

Type Definitions

ConfigKeyMap