Module dprint_core::formatting

Modules

condition_resolvers

conditions

parser_helpers

tokens

utils

Structs

Condition

Conditionally print items based on a condition.

ConditionProperties

Properties for the condition.

ConditionReference

ConditionResolverContext

Context used when resolving a condition.

Info

Can be used to get information at a certain location being printed. These can be resolved by providing the info object to a condition context’s get_resolved_info(&info) method.

PrintItems

Print Items

PrintItemsIterator

PrintNode

Print Node

PrintNodeCell

A fast implementation of RefCell that avoids runtime checks on borrows.

PrintOptions

Options for printing the print items.

StringContainer

A container that holds the string’s value and character count.

WriterInfo

Information about a certain location being printed.

Enums

PrintItem

The different items the printer could encounter.

Signal

WriteItem

Functions

format

Function to create the provided print items and print them out as a string.

print

Prints out the print items using the provided options.

Type Definitions

ConditionResolver

Function used to resolve a condition.

PrintItemPath