Module dprint_core::formatting::parser_helpers

Structs

LinesSpan

MultiLineOptions

ParseSeparatedValuesOptions

ParseSeparatedValuesResult

ParsedValue

Enums

BoolOrCondition

Functions

new_line_group

parse_js_like_comment_line

parse_raw_string
Parses a string as is and ignores its indent.

parse_separated_values

parse_string
Parses a string to a series of PrintItems.

surround_with_new_lines

surround_with_newlines_indented_if_multi_line
Surrounds the items with newlines and indentation if its on multiple lines. Note: This currently inserts a possible newline at the start, but that might change or be made conditional in the future.

text_has_dprint_ignore
Gets if the provided text has the provided searching text in it (ex. “dprint-ignore”).

with_indent

with_indent_times

with_no_new_lines

with_queued_indent